awsの設定ファイルのパスの変更
文字列結合、ファイルパスのリファクタリング

+++ src/app copy.py
@@ -1,13 +1,7 @@
 import os  # オペレーティングシステム関連
 
-# 初期化
-error_log = None
-
-# エラーログのインポート
 try:
     from package.error_log import ErrorLog  # エラーログに関するクラス
-    # エラーログ作成
-    error_log = ErrorLog.create_error_log()
 # インポートに失敗したなら
 except Exception as e:
     message = [
@@ -17,32 +11,7 @@ except Exception as e:
         "管理者に問題を報告していただけると幸いです。",
     ]
     print("\n".join(message))
-    exit() # 処理を終了する
 
-
-# その他の自作クラスのインポート
-try:
-
-    from package.fn import Fn  # 自作関数クラス
-
-    from package.user_setting import UserSetting  # ユーザーが変更可能の設定クラス
-    from package.system_setting import SystemSetting  # ユーザーが変更不可の設定クラス
-
-    from package.window.translation_win import TranslationWin  # 翻訳画面ウィンドウクラス
-    from package.window.display_setting_win import DisplaySettingWin  # 表示設定画面ウィンドウクラス
-
-    # 環境設定画面ウィンドウクラス
-    from package.window.environment_setting_win import EnvironmentSettingWin
-    from package.window.key_setting_win import KeySettingWin  # キー設定画面ウィンドウクラス
-    from package.window.language_setting_win import LanguageSettingWin  # 言語設定画面ウィンドウクラス
-    from package.window.save_setting_win import SaveSettingWin  # 保存設定画面ウィンドウクラス
-    from package.window.shooting_setting_win import ShootingSettingWin  # 撮影設定画面ウィンドウクラス
-    from package.window.theme_setting_win import ThemeSettingWin  # テーマ設定画面ウィンドウクラス
-    from package.window.user_info_win import UserInfoWin  # 利用者情報画面ウィンドウクラス
-
-except Exception as e:
-    # エラーログの出力処理
-    ErrorLog.output_error_log(error_log, e)
 
 class App:
     """アプリケーションのメインクラス"""
@@ -56,6 +25,22 @@ class App:
     def run(self):
         """メインの処理"""
         # 必要クラスのインポート
+        from package.fn import Fn  # 自作関数クラス
+
+        from package.user_setting import UserSetting  # ユーザーが変更可能の設定クラス
+        from package.system_setting import SystemSetting  # ユーザーが変更不可の設定クラス
+
+        from package.window.translation_win import TranslationWin  # 翻訳画面ウィンドウクラス
+        from package.window.display_setting_win import DisplaySettingWin  # 表示設定画面ウィンドウクラス
+
+        # 環境設定画面ウィンドウクラス
+        from package.window.environment_setting_win import EnvironmentSettingWin
+        from package.window.key_setting_win import KeySettingWin  # キー設定画面ウィンドウクラス
+        from package.window.language_setting_win import LanguageSettingWin  # 言語設定画面ウィンドウクラス
+        from package.window.save_setting_win import SaveSettingWin  # 保存設定画面ウィンドウクラス
+        from package.window.shooting_setting_win import ShootingSettingWin  # 撮影設定画面ウィンドウクラス
+        from package.window.theme_setting_win import ThemeSettingWin  # テーマ設定画面ウィンドウクラス
+        from package.window.user_info_win import UserInfoWin  # 利用者情報画面ウィンドウクラス
 
         # ウィンドウクラスのマッピング辞書
         WIN_CLASS_DICT = {
@@ -80,16 +65,9 @@ class App:
         # ユーザ設定のインスタンス化
         user_setting = UserSetting()
 
-        # # AWSサービスにアクセス可能か確認する処理
-        # aws_service_exception = user_setting.check_access_aws_service()
+        # AWSサービスにアクセス可能か確認する処理
+        user_setting.check_access_aws_service()
 
-        # # AWSサービスにアクセス時に発生した例外オブジェクトが存在するなら
-        # if aws_service_exception is not None:
-        #     print("AWSアクセスエラー")
-        #     print(aws_service_exception)
-        #     return
-        # else:
-        #     print("正常")
 
         # 翻訳前、後画像の両方が存在しない履歴ファイルを削除
         Fn.delete_unique_history_file()
@@ -107,5 +85,8 @@ class App:
             transition_target_win = win_instance.get_transition_target_win()  # 遷移先ウィンドウ名取得
         Fn.time_log("システム終了")
 
+
 if __name__ == "__main__":
     app_instance = App()  # メイン処理
+    # print(app_instance)
+    from package.global_status import GlobalStatus  # グローバル変数保存用のクラス
